# nestjs-backend/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/portfolio/{id}": {
      "get": {
        "operationId": "PortfolioController_getByPortfolioId",
        "summary": "Get one Portfolio by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID of the portfolio to fetch",
            "example": "portfolio123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The portfolio with the specified ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioDto"
                }
              }
            }
          }
        },
        "tags": [
          "Portfolio Service"
        ]
      }
    },
...
    "/core/{manager}": {
      "get": {
        "operationId": "CoreController_loadHomepage",
        "summary": "Get all Portfolios by Manager",
        "parameters": [
          {
            "name": "manager",
            "required": true,
            "in": "path",
            "description": "Manager ID",
            "example": "66d9815bacb3da812c4e4c5b",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all portfolios under specific manager id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "portfolioId": "66d9ae695e15ad24b5e2053a",
                        "clientName": "Tom",
                        "portfolioName": "Portfolio1",
                        "riskAppetite": "MEDIUM",
                        "totalValue": 15366.77,
                        "totalPL": 1921.6,
                        "dailyPL": -11.63,
                        "totalPLPercentage": 14.36,
                        "dailyPLPercentage": -0.08,
                        "rateOfReturn": 100,
                        "alertsPresent": true
                      },
                      {
                        "portfolioId": "673edcfaa141dc2133e204e6",
                        "clientName": "IS484",
                        "portfolioName": "Final Presentation",
                        "riskAppetite": "MEDIUM",
                        "totalValue": 119342.5,
                        "totalPL": -657.5,
                        "dailyPL": -102.19,
                        "totalPLPercentage": -0.69,
                        "dailyPLPercentage": -0.11,
                        "rateOfReturn": 100,
                        "alertsPresent": true
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
    "/core/portfolio/{portfolioId}": {
      "get": {
        "operationId": "CoreController_loadPortfolio",
        "summary": "Load Portfolio data of a specified portfolio",
        "parameters": [
          {
            "name": "portfolioId",
            "required": true,
            "in": "path",
            "description": "portfolio Id of portfolio",
            "example": "66d9ae695e15ad24b5e2053a",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Portfolio details by portfolio",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "portfolioId": "673edcfaa141dc2133e204e6",
                    "clientName": "IS484",
                    "portfolioName": "Final Presentation",
                    "portfolioAnalysis": {
                      "totalAssets": 119342.5,
                      "cashAmount": 24500.000000000022,
                      "securitiesValue": 94841.48474999999,
                      "dailyPL": -102.19,
                      "dailyPLPercentage": -0.11,
                      "totalPL": -657.5,
                      "totalPLPercentage": -0.69,
                      "annualizedRoR": 100
                    },
                    "triggeredAlerts": [
                      {
                        "id": "673dc25e23fe016af3c6fba5",
                        "ticker": "CAT",
                        "date": "2024-11-20T19:05:02.920Z",
                        "sentimentRating": 4,
                        "introduction": "Caterpillar...",
                        "assetName": "Caterpillar"
                      }
                    ],
                    "breachedRules": {
                      "breachedRules": [],
                      "recommendation": "",
                      "news": {
                        "buy": []
                      }
                    },
                    "portfolioBreakdown": {
                      "securities": [
                        {
                          "STOCK": 39.68
                        },
                        {
                          "BOND": 39.79
                        },
                        {
                          "CASH": 20.53
                        }
                      ],
                      "industry": [
                        {
                          "Construction and mining": 3.58
                        }
                      ],
                      "geography": [
                        {
                          "USA": 100
                        }
                      ]
                    },
                    "portfolioHoldings": [
                      {
                        "name": "Caterpillar",
                        "ticker": "CAT",
                        "type": "STOCK",
                        "geography": "USA",
                        "position": 9.021920934257132,
                        "market": 3394.046765598404,
                        "last": 376.20001220703125,
                        "cost": 387.510009765625,
                        "totalPL": -102.037903740274,
                        "totalPLPercentage": -2.92,
                        "dailyPL": -27.426441404574277,
                        "dailyPLPercentage": -0.8,
                        "positionsRatio": 3.58
                      }
                    ],
                    "orderExecutionProgress": []
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
...
    "/core/news/{id}": {
      "get": {
        "operationId": "CoreController_loadFinanceNewsById",
        "summary": "News by id for individaul article page",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Individual finance news by ticker id",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "66ef29f6d094c73406fa5ea2",
                    "company": "Apple",
                    "ticker": "AAPL",
                    "date": "2024-09-22T04:17:58.562Z",
                    "sentimentRating": 4,
                    "introduction": "As of September..."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
    "/core/ruleLogs/{portfolioId}": {
      "get": {
        "operationId": "CoreController_loadRuleLogs",
        "summary": "Get all Rule Logs by portfolioId",
        "parameters": [
          {
            "name": "portfolioId",
            "required": true,
            "in": "path",
            "description": "portfolio Id of portfolio",
            "example": "66d9ae695e15ad24b5e2053a",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleLogDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
    "/core/portfolio/validate/{portfolioId}": {
      "post": {
        "operationId": "CoreController_validateIntermediatePortfolio",
        "summary": "Validate intermediate portfolio rules",
        "parameters": [
          {
            "name": "portfolioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Validate intermediate portfolio rules by providing asset holdings and cash amount",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateIntermediatePortfolioClass"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Core Service"
        ]
      }
    },
...
    "/financeNews/{ticker}": {
      "get": {
        "operationId": "FinanceNewsController_getByTicker",
        "summary": "Get one Finance News by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns specific finance news by ticker",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "66ef29f6d094c73406fa5ea2",
                    "ticker": "AAPL",
                    "date": "2024-09-22T04:17:58.562Z",
                    "sentimentRating": 4,
                    "summary": [],
                    "references": []
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Finance News Service"
        ]
      }
    },
    "/financeNews/tickers/latest": {
      "get": {
        "operationId": "FinanceNewsController_getLatestByTicker",
        "summary": "Get the latest news for all specified tickers",
        "parameters": [
          {
            "name": "tickers",
            "required": true,
            "in": "query",
            "description": "Array of ticker symbols to fetch finance news for",
            "example": [
              "AAPL",
              "DOW",
              "AMGN"
            ],
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns latest finance news of specific tickers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "66ef29f6d094c73406fa5ea2",
                        "ticker": "AAPL",
                        "date": "2024-09-22T04:17:58.562Z",
                        "sentimentRating": 4,
                        "summary": [],
                        "references": []
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Finance News Service"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio Optimizer Application API",
    "description": "core application in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AssetHoldingDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "assetType": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "cost",
          "quantity",
          "assetType"
        ]
      },
      "PortfolioDto": {
        "type": "object",
        "properties": {
          "client": {
            "type": "string"
          },
          "portfolioName": {
            "type": "string"
          },
          "riskAppetite": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "type": "string"
          },
          "cashAmount": {
            "type": "number"
          },
          "assetHoldings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetHoldingDto"
            }
          },
          "manager": {
            "type": "string"
          },
          "exclusions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rules": {
            "type": "object"
          }
        },
        "required": [
          "client",
          "portfolioName",
          "riskAppetite",
          "cashAmount",
          "assetHoldings",
          "manager",
          "exclusions",
          "rules"
        ]
      },
...
      "RuleLogDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "managerId": {
            "type": "string"
          },
          "portfolioId": {
            "type": "string",
            "nullable": true
          },
          "ruleType": {
            "enum": [
              "MIN_CASH",
              "MAX_CASH",
              "RISK",
              "EXCLUSIONS"
            ],
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          },
          "changeMessage": {
            "type": "string"
          }
        },
        "required": [
          "description",
          "managerId",
          "portfolioId",
          "ruleType",
          "timestamp",
          "changeMessage"
        ]
      },
      "ValidateIntermediatePortfolioClass": {
        "type": "object",
        "properties": {
          "intermediateAssetHoldings": {
            "description": "Intermediate asset holdings to be validated",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "intermediateCashAmount": {
            "type": "number",
            "description": "Intermediate cash amount to be validated"
          }
        },
        "required": [
          "intermediateAssetHoldings",
          "intermediateCashAmount"
        ]
      },
...
    }
  }
}